requirements.txt
requests==2.26.0
APScheduler==3.7.0
python-dotenv==0.19.0